modbus: Prevent panic with registers more than max-register apart
Modbus supports querying 2000 digital registers or 125 analog registers
at a time. If a user would configure multiple e.g. analog registers,
that can not be quieried with a single query (=> not within 125
distance), then the exporter would try to access indices in the returned
byte slice bigger than the max-register size (125).

This patch:

- Makes sure to only access indices within the max register size (2000 /
125).

- Breaks the loop once all registers have been found.

- Adds test framework as well as two tests for `getModbusData`.

- Adds project binary to `.gitignore`.