*: Introduce data types
Modbus operates on 16 bit registers. These opaque bit streams can be
interpreted as different data types. This patch adds support for the
basic ones to be configured per metric.

In order to achieve the above, this patch also includes a couple
refactorings:

- Remove lexer and parser, instead configure the exporter with idiomatic
YAML, without a custom language on top. For example:

  ```
  targets:
    - name: simulator
      port: 127.0.0.1:1502
      id: 1
      analogIn:
        - name: register_twenty_two
          address: 22
          dataType: int16
  ```

- Rename `config.Slave` to `config.Target`.