modbus: Only request two registers per metric definition
As of now, the exporter does a single request per metric definition.
Thereby requesting the maximum amount of registers per request is
unnecessary and error prone in case it falls over the edge of overall
maximum available registers.

Instead, until proper caching is implemented, with this patch only two
registers are requested. Two due to the fact that the largest data type
is a float32 which spanns two registers.