*: Configure modbus unit aka. sub target in prometheus.yml
Instead of configuring the modbus unit per modbus module in the
modbus.yml configuration file, with this commit the modbus unit is
configured in prometheus.yml and passed to the exporter via an http
parameter.

This keeps the modbus exporter configuration file (modbus.yml) generic
(containing only hardware specific module definitions) and the modbus
unit (0-255) part of the information that identifies a Prometheus scrape
instance.