The error messages of the `InsufficientRegistersError` was misleading since it compared number of registers (aka 16bits) with number of bytes of raw data. For example an erroneous configuration that sets the data type to `int16` where a `bool` would have been correct, yields the confusing message:
```
insufficient amount of registers provided: expected 1, got 1
```

This fixes this issue by printing the number of expected bytes not registers. Using the
number of registers on both sides of the comparison would be more in line with the name
of the error but would have lead to awkward messages like

```
insufficient amount of registers provided: expected 1, got 0.5
```
Imho using bytes is easier to understand.

Signed-off-by: Christian Pointner <equinox@spreadspace.org>